backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
